GitStatusMonitor could start multiple commands (#9699)
If the monitor was inactivated while a status command was running
(minimizing/restoring the app, start a modal dialog) the current command
is canceled to give the latest data.
lock() will ensure that more than one "active" command is not running.
However, one control variable was reset also if the command was canceled
which allowed the canceled command to be followed by another.

Also changed to not request a new status command if a command is already
running. (Current status should be good enough, if there are pending
changes they will be requested later.)

Cherry picked essential changes from master:
#9700/e2630a6c5b09452b0a0ef7a3ff169083c3fcb7c4