PathFilter: Set FallbackFollowedFile at every filter update
* Adjust the paths for FallbackFollowedFile by removing quotes.
As before, user specific entries with wildcards and multiple files cannot
be handled.

* Update FallbackFollowedFile in RevDiff and FileHistory also when changed
in the UI filter (not just from the tab menus).

* Do not update filters by nullifying the pathfilters when switching repos.
This triggers additional FilterChanged events.
As a bonus, the pathfilter info is available (disabled) in the filter.
Could be useful if switching to a new worktree or when switching back to the
old repo.