Replace AppSettings.UseFastChecks with ShowGitStatusInBrowseToolbar
Remove the specific setting for the index watcher, use the GitStatusMonitor
settings instead. UseFastChecks hardly adds any overhead, it adds
FileSystemWatcher to index and refs (GitStatusMonitor has a separate
watcher on index).
If the file system is changed, the top-left refresh button becomes red.
(As well as avoiding refreshing the grid in some situations, may not be working though.)

UseFastChecks was visible in Performance settings and expected by
users (and occasionally maintainers) to make a difference.
Sometimes the setting is mixed up with the GitStatusMonitor setting
that can make a difference (as git-status requires a lot of resources,
not FileSystemWatcher).