AdvancedFilter: Reset filters button dropdown
Change Advanced filter to a split button and add "Reset filters"
to the dropdown for all and path filters.

* To indicate that a filter is active, the "Checked" property can no longer be used.
This is indicated with a new icon instead.

* Update hotkey for filter related actions

Ctrl-F is also used for search, use Ctrl-I instead.
Similar for disabling filters (added in master),
Ctrl-Shift-F is used for searching files and is adjusted to
Ctrl-Shift-I.
This PR adds a hotkey for disabling the path filter,
set to Ctrl-Shift-H.