BuildServer: Unset settings were always set to effective (9785)
Setting when editing a specific level (Local<-Distributed<-User (global)).

Checkboxes now reflects if the bool value is set on the specific level.

For textboxes will an empty string be handled as unset (i.e. null).
Without this will lower priority values always have to be set in
higher priority settings. Order is Local<-Distributed<-User (global).

For BuildServer settings where an empty string could make sense to
override a value (like ignore a setting in Distributed by setting it locally)
adummy value must be used.
This is still considered a better functionality than requiring the value to be set.

* Displayed settings were not updated when changing level

* Add unit and integration tests