Blame: Start on line also when control already exists (#9832)
Blame can start presenting on a specific line. This is is available in
RevDiff and RevFileTree since #9424.
(Even if the code enabling this has existed for long time,
it was just used for a special case.)

The "goto line" was only used if the control had to be initialised,
not if e.g. toggling to Diff and selecting an new diff line
and then Blame again.