Offload refs and revision loading to background thread
This is a tactical fix aim at increasing the responsiveness of the app
during the refresh by moving Module.GetRefs call and revisionReader.Execute
calls to a background thread.
This makes the app responsive even while loading huge repos
like VS, which has >100K commits and takes more than 10s to parse
revisions info, and has close to 180K tags, which also take more than 10s
to parse.
NB: It also takes another 10+s to load that many tags into the left
panel, and this issue is not being addressed in the context of this change.

A more holistic fix would likely involve rework of Module.GetRefs implementation.

Also pass a cancellation token to allow cancel the load operation.

Lastly show the spinner while revisions are being loaded, and clear both
the left panel and the commit info panel at the beginning of the refresh.