FileHistory mode at browse startup
When starting up with command line arguments "filehistory", "blamehistory"
(for instance from VS or explorer) or "browse <path>" (file history in the GUI),
hide the side panel and start in FileTree tab with Blame active to
provide an experience similar to old the FileHistory.

The forced side panel state is not persisted to settings at shutdown.