IndexWatcher: Use for refresh icon only
The IndexWatcher controls the icon for the refresh button in Browse.
Before #9791 the setting could be controlled individually,
now the same setting as for GitStatus (commit count).

Previously it also tried to optimize grid refresh operations in a few
situations, skipping when no changes was detected to the index or any refs.
These checks are now removed, always refreshing the grid.

The check for changes seem to miss updates occasionally (and never in
WSL where FileSystemWatcher do not reeport updates).
This seem to be improved by matching on creation of index.lock
in addition to index, but then also reset of worktree files
are trcked as index changes.

Before the file watcher was initiated at creation of the control and
when a new revision was loaded. This requires a sync call to
git-rev-parse --git-common-dir delaying the revision loading
50-100 ms.
This is changed to just clear the icon when loading a revision
(as well as when a new repo is loaded even if that should not be required)
and resetting the file system watcher after all revisions was loaded.

The IndexWatcher was reset several additional times after the grid
was refreshed.

The IndexWatcher.IndexChanged was used to possibly refresh the revisions
that was already refreshed when the revisions were selected in the grid.
(Also, no changes should have occurred this time.)