Browse: Clarify RefreshRevisions() handling
If a new repo is opened in FormBrowse the call chain is
SetGitModule() -> UICommands_PostRepositoryChanged() -> RefreshRevisions()
Most other refreshes in Browse uses RepoChangedNotifier.Notify(),
which is confusing when reading the code.

As the sidepanel is init from RepoChangedNotifier.Notify(),
the use of Notify() cannot be changed now.
This commit just separates the changes to RefreshRevisions().