RefreshGrid: async reading refs and logs in parallel (#9864)
Improve the responsiveness in the loading phase and decrease
the revision loading time.

Revision info is read in two parallel steps:

1. Read current commit, refs, prepare grid etc.
2. Read all revisions with git-log.

Both these operations can take a long time and is done async.
Before part 1. was read synchrounously and blocked the interface
(Part 2 (git-log) was async).
As the information in step 1. is needed by step 2. when commits are added to the
grid, the information from step 1. is therefore protected with a semaphore.

RevisionReader (git-log) was previously also updating other information,
but is now a pure revision reader.

Other:

* Share GetRef() calls also when initiating from RevGrid.
Still some redundant calls related to the sidepanel and no branch checked out to be handled separately.

* The text label informing that loading of revisions still occur was
not displayed.

* NavigationHistory was normally always reset also if the grid was refreshed.
(With a potential race condition to keep existing history also if a new repo was
opened).

* Avoid a Git call to get current head.

* Optimizations for git-notes, assuming no notes exists.

* Avoid refreshing tab contents an extra time when adding or removing tabs

Co-authored with russkie@gmail.com

Co-authored-by: Gerhard Olsson <gerhardol@users.noreply.github.com>