fixup! Load grid to selected revisions
The grid loads async separately from the revisions, must be awaited
before selecting rows in the grid or executing tests.
If the revisions are 'recent' the rows will normally be loaded, but not if
few revisions were loaded (like in tests) or an early commit were
selected in another repo.

If only some of the revisions were found, show those.
`CheckAndRepairInitialRevision` that tried to handle this before were
not running correctly.

If HEAD was not seen for artificial commits or the previously selected
revision was filtered, try to find a parent instead.

Handle grids without HEAD where artificial commits are inserted.
Previously, the selection "shifted" at refresh.

Abort 'restore' selection during load if manual selection is performed.

GridLoaded event renamed to RevisionsLoaded to reflect usage.