Show stashes in the revision grid
Show stash commits in the revision grid, with menu options to apply,
pop and drop stashes.

Stashes are listed in parallel to refs and the first part reading logs,
to not delay the revision display in the grid.
The first stash is always shown similar to current implementation,
other stashes are shown only if the parent commit is shown
(so all stashes are chown with rev grid).

(Except for the first grid and for reflog, the index commit is not shown.
The internal commit with untracked files is hidden if it is empty for
the first stashes. (Limited to not delay the revision display.)
Further performance evaluation in the pull request.