squash! No revision filter search when changing filter selection
Handle the case where:
1. Search for a string (by default search in commit message)
2. Select "Diff contains" and unselect "Commit message"
3. Launch a search but search/filtering is not launched

* In 3.5, a revision filter search were done for all checked menu items.
In master only the first checked was used, in the following order:
Author, Committer, Message, Diff

In general, a user wants to use one of them, multiple options can be
search in the Advanced filter form.
The behavior was therefore changed to only select one of the options.

* ByDiffContent was not part of the criteria for "HasFilter".
Similar should be done for the following, but they should be included in
Reset/Summary first:
ShowCurrentBranchOnly, ShowReflogReferences, ShowFirstParent.
(AppSettings.ShowMergeCommits and other handled as settings should
probably be excluded.)
These fields are not visible in the Advanced form though.

* Remove unused menu item for hash