Set branchfilter from sidepanel without checking dropdown
The branchfilters from the sidepanel was not applied unless the
filter dropdown contained any items, but the items were not updated
by default since #10136, only when needed (dropdown clicked or text typed).

The dropdown was updated async FileAndForget, so 'external' usage was not
always successful. This was a reason why the test
'Filters_should_behave_as_expected' had to be disabled in AppVeyor
(the test still occasionally fails though).