Blame: Fix path displayed in the gutter (#10188)
It should be displayed only when the path is different from the current file blamed
but the comparison failed because the relative path passed is in Windows format
whereas the ones of blame lines are in Unix format.

So converting the one passed into Unix format for comparison.

(cherry picked from commit fc90399de90473735183c15ce6e837168c7a98c5)