Move revision filters to advanced filter form
Most revision filters are available in the Advanced filter form
(FormRevisionFilter), some are in View menu and some in the toolbar.
This PR adds all revision filters to FilterInfo and
FormRevisionFilter and removes the specific from View
menu (except Branch filters, that is still available in View
and the toolbar in addition to Advanced form).

Concentrate the logic for revision filters to FilterInfo,
some handling were previously in RevisionReader too.

ShowSimplifyByDecoration were not included for
ParentsAreRewritten.

Consistently rnamed ShowFirstParent to
ShowOnlyFirstParent.

Added mnemonics and Enter handling.

Tests are adapted and slightly expanded.

Reorder controls in the form in order they are placed,
to adjust the generated indexes.