Reorder and group items in View menu (#10189)
Group the menu items to reflect how the items are used,
separated with group headings

A number of related fixes:

* stash and notes could be shown with current and filtered branches
This normally did not occur as they were filtered, but it affected tests.

* Decouple ShowMergeCommits and ShowRevisionGridGraphColumn
(column can be shown when this filter is enabled)

* Reflog was not coupled with other branch revision filters
Even if Reflog dominated the other branch revision filters,
this was not reflected in the menus.
Also toggling the reflog (with the button) would set the branchfilter
to all branches.

* Add reflog to branch filter dropdown, to show current filter

* Consistently name reflog/reflogs to "reflog"
Except for "noreflogs" for "fsck-objects --no-reflogs"

(cherry picked from commit 480458b30b30c883ec1a9c07d571464233cdb319)