Restore the order of the branch view options
* The vast majority of users would likely use "all branches" most of the
time because it's the default option and because it's the git's default.

* Filtering - i.e., viewing the current branch or selected other branches
- would likely be the 2nd and 3rd most used options.
The order has been like this for at least a decade, so users likely havedeveloped muscle memory.

* Reflogs is likely the least known feature to many git users, and to
promote this feature more a button was added to the toolbar in #9528.
But since it's not likely used by the majority of users, it's inappropriate
to have it placed before the other items.