Custom options to git-log
If "Message" starts with "--not " or "--exclude=", interpret the string
as options to git-log.
This enables the possibility to enter filters that the GUI
does not support (like negating conditions).

The options has no checks for Git syntax or if ParentsAreRewritten etc
If need to be adjusted, it is up to the user.