Custom options to git-log (#10240)
If "Message" starts with "--not " or "--exclude=", interpret the string
as options to git-log.
This enables the possibility to enter filters that the GUI
does not support (like negating conditions).

The options has no checks for Git syntax or if ParentsAreRewritten etc
If need to be adjusted, it is up to the user.

(cherry picked from commit e703d0c65e71ee1ac49bec57f0aa4a08cc5058b4)