Cleanup handling of AllOutput for processes (#10262)
AllOutput is combined use of stderr and stdout and should only
be used to display messages for users.

* Change a few commands to explicitly read stderr where required.
Some tests had to be adapted (including MockExecutable).

* Clarify usage that AllOutput is used for user popups.
Most of these occurrences are related to FormCommit.
The structure of these command was changed to return a bool for
result and the output as a string, also if result/output were unused.

* Set TODO on a few occurrences that were not investigated.

(cherry picked from commit 2e2d79f870173dee349615887f581a0b04a1f1aa)