Fix 'Delete obsolete branches' plugin
When adding nullable reference types to this code, an incorrect assertion was added. The nullability of the cancellation object here is quite confusing, as setting IsRefreshing to true causes it to be created. It's also known as not-null if IsRefreshing is true. This commit updates the assertions to be correct, fixing the bug.