Retain only Microsoft.WindowsDesktop.App in *.runtimeconfig.json (#10355)
Resolves gitextensions#10337

SDK currently puts the lower level framework first (in runtimeconfig.json),
so the hosts sees that first and if it's missing it fails - so that's why
it produces the missing Microsoft.NETCore.App message - and the related
URL points to that download.

If there are two frameworks listed in runtimeconfig.json with any of them
missing, it's impossible for the host to determine which is lower level
and which is higher level.

To avoid the issue, only retain the highest level framework - which is
the WindowsDesktop, so users see correct message and get the correct
link to download the required Runtime.

(cherry picked from commit 554f88459590566139aa8d7ac9c45b39196f0cee)