Fix 'Delete obsolete branches' plugin (#10396)
When adding nullable reference types to this code, an incorrect assertion was added.
The nullability of the cancellation object here is quite confusing, as setting IsRefreshing
to true causes it to be created. It's also known as not-null if IsRefreshing is true.
This commit updates the assertions to be correct, fixing the bug.

(cherry picked from commit 1d1afd58cf8322a746f9fc559e4605c8415d2a3e)