Update README.md
Removes dragons because they are not real and have no basis in nature. Have you ever seen a dragon skeleton? There is no way a dragon can support its weight and there is no way the bones for the wings fit into the normal skeletal pattern of actual animals that exist.