RefreshGrid without revisions
If option "Show only current branch" was set, the repo was required to have
commits and a default branch for git-log not to report errors.
(No revision arguments to git-log by default.)

In addition, if stashes were shown the first stash to show requires
that a branch was checked out (by explicitly specifying HEAD).

To avoid these scenarios, a branch filter is always supplied to
git-log, even if no revision arguments to Git is the defaults.