Do not delay Dashboard start for invalid Git repos
Add the repos to the lists without any I/O to check
if the repo is invalid or not (to set the icon).
Perform the check if a repo has a valid Git dir async,
together with branchname retrieval and set the invalid icon
then instead.

Rendering of branchname failed if the item was rendered before
the branchname was set.
This had a much higher probability to occur after the
item creation was changed.

Menu item to remove invalid repos was only set from
favorite repos, as _hasInvalidRepos was overwritten.