ShowCurrentBranchOnly: Filter also if no branch checkout (#10944)
Fixes #10939

If no branch checkout, the branch filter was handling ShowCurrentBranchOnly
as ALBranches.
This is changed to work similar to before, where ShowCurrentBranchOnly
filtered on the current commit instead.
A consequence is that the first stash is not added to the grid if there is no
branch checked out.

(cherry picked from commit b148979eddcab81cde6fbd661c1377e430cbae05)