RevDiff: Reset to HEAD for artificial
Instead of resetting to Index, reset to HEAD when to artificial revisions.
Reset to Index does not mean anything.
This aligns to resetting in e.g. FormCommit and normal Git behavior.

Note: Reverse clicking (artificial before normal commit) will now list
the normal commit as "first" and assign the hotkey.