Reset files corrections
Use same code in RevDiff as in FormDiff and command line to reset
individual files.

Reset index explicitly before resetting to commits.
(This was done for FormCommit Index reset before.)
This requires additional calls to git-status.

Similar, files that are new in the worktree is checked with git-status,
so new files are deleted properly.

This handles some more scenarios for unmerged (confliced) and renamed files.

Also, criteria to allow delete new file in RevDiff is more generic.