Prevent a (not important) early git call breaks the ownership security issue handling done in GE
because this early git call doesn't capture the output
so the repo ownership security issue handling can't detect that the crash is due to this reason
and so display the crash popup instead of the user friendly ownership issue dialog.

Because these 2 actions are not very important and very unlikely to crash, swallowing the exceptions
to let subsequent git calls crash if needed and handle the security issue well.

Fixes #11039 (and #10975 )