Blame: better default value for "change move detection"
because when enabled the blame does not correspond **exactly** to the diff on file done
and it could be very misleading and the user could spend a lot of time trying to understand
why a line is displayed to be introduced by a commit but it doesn't see it in the diff!
So the option should be carefully and purposefully set by the user.

This default option is not really compatible with investigations where
the "Blame this revision" or "Blame previous revision" feature are used
**on detected moved code**
because user end up in the wrong place.

Ex: in GitExtension repo,
 on commit e889d7513e9bf8eb98c3d6a9b62c29c3dd8ccbc2
 the line 2263
 of file "GitUI/CommandsDialogs/FormBrowse.cs"
 is falsly displayed as introduced by commit 52476f30670ba5338756b606841fb0a346fd6460
 instead of really be introduced by e889d7513e9bf8eb98c3d6a9b62c29c3dd8ccbc2 (see diff)

Already discussed in https://github.com/gitextensions/gitextensions/pull/4454#issuecomment-364856434

+ nit: in settings file, move line to be in same order than GUI