Rebase: improve perfs when displaying list of commits
by retrieving all the commits data with only one `git log` call
instead of one by commit.
It is possible as when rebasing, all the commits belong to the same branch.

Kept a fall back to previous way to retrieve commit data for 1 commit
if commit is not in the one loaded.

and read bodies of all rebased commits
(even the ones older than 6 monthes like it was the case before.
RevisionReader had to be adapted to do that.)

Results with 7448 commits rebased: 2s instead of 4m50 previously

Fixes #10698 and #11196