rebase: fix accuracy of data in list of commits displayed
by using "done" file for commits already processed
and "git-rebase-todo" for what will be **really** done.

because "git-rebase-todo.backup" content is not accurate
(it contains what was presented to user in interactive rebase
**before** user choose which actions to do on commits)

it fixes "action" column value that was not accurate
and commits that was reordered.

Note: for commits that we don't want to keep, the normal behavior is the following:
* 'dropped' commits are still displayed but with the action 'dropped'
* removed or commented commits lines are not displayed in the list
So prefer using 'd'/'dropped'...

fixes #11179