Source generate the boilerplate
Windows Forms designer requires a default constructor to be able
to open a form on the design surface.

This, however, does not work in situations where some parameters must be
set in the constructor (e.g., dependency injection).

Previously, we would create a default constructor specifically for the
Windows Forms designer integration. However, this implies a lot of
unnecessary overhead.

Emit the default constructors for the designer integration with source
generators.