Use RevisionReader to read GitRevision with Notes
Do not allocate when replacing VerticalFeed.
Add vertical feed test.

Optionally read Git Notes in RevisionReader
Tests were added.

Use RevisionReader to parse revisions also in
CommitData and GitModule.
This require support of Notes.

Add LogOutputEncoding to IGitModule so RevisionReader can use the
interface internally.

If showing Git Notes, read them when loading the grid so they are handled
with the Body in the standard scenario.