RevisionGrid: Separate add/insert of artificial commits
If HEAD is not included in git-log, the artificial commits are
inserted after all commits are loaded.

This is a refactoring to separate the normal add and the insert to
simplify the handling.

Set the score where to invalidate the node cache from
and check the parent scores only once.