Limit HEAD parents when inserting artificial in grid (#11266)
HEAD may not be included in git-log in some scenarios.
One example is an old checkout in the Linux repo where 100K commits
is about a year. Another is when filtering.

In this scenario there is no HEAD the artificial commits
cannot be inserted while loading.
The HEAD parents are then listed to find the first descendant in the grid,
or first in the grid if no head parent is found.

Similar scenario for the revision to be reselected.
When switching repo the HEAD is always selected.

Previously, a match was tried in all HEAD parents.
This command requires 1.5 s in the linux repo,
delaying the the time until the spinner is dismissed.
The number of parents to try is limited to around 50 ms instead.

In addition, the command ran a second time as the
result was not enumerated when trying to find the revision to select.