RevisionGraph: improve internal structures
General changes to improve cpu execution and to
limit allocations. Some changes are very minor like
using for instead of foreach, other changes the data structures.

Separate storage of nodes that has no GitRevision.
This allows to remove a separate node storage and
to simplify the code.

Use ImmutableArray instead of array and list for the ordered caches
to improve lock handling.

Force update data grid when revision graph nodes have been modified

The order of nodes in the revision graph may change if the commits are
listed out of order or if artificial commits are inserted.
The revision graph was (normally) invalidated but the data grid was only
refreshed if scrolling.

Use separate counter for revisions to limit
the lock time for _revisionByObjectId.Count

Rename variables and add comments

Remove unused IsRevisionRelative().

and a merge error in tests