perf(plugins): loading only IRepositoryHostPlugin when opening on dashboard (#11538)
to make "Clone GitHub repository" appears quickly
(and not after multiple seconds)
and then load other plugins after.

(Not so) BREAKING CHANGE: a git hosting plugin need to export `IRepositoryHostPlugin`
Not really a breaking change because the rest of the application
actually doesn't support a 2nd one and no one ever complained.
So that a clue that there should be no other git hosting plugin that has ever been developped...