FormCommit: fix registering plugins timing regarding to the trigger events (#11538)
(i.e for FormCommit standalone window)

because the PreEvent (i.e PreCommit) event was triggered **before** plugins were loaded and registered
and postEvent (i.e PostCommit) was triggered after the plugins were unregistered

So plugins were still not loaded while PreCommit is triggered
and some (commit message helpers) plugins are subscribing to this event
and so should be loaded before.

Looking at `DoActionOnRepo` method could help understand the problem. It is behaving like that (pseudo code!):

```
DoActionOnRepo()
{
    PreCommit();
    Action();
    PostCommit();
}
```

So, if plugins are loaded in `Action` lambda passed, it could not work...