git-grep in Diff tab
Search commits (as well as worktree/index) without checking out those commits.

A "global search" popup with shortcut Ctrl-Shift-F as in many IDE.
It is also possible to add a searchbox to the Diff tab,
configurable in the search popup menu.

If there is no `-e ` in the search string, this is added to the search string.
Search options like `--and` may be added.

It is possible to add persistent search options in a separate box in the
popup (there are checkboxes for the most common case insensitive).
Search options like `--show-function` may be set as a preference for a user