amend! RevDiff Find file clarification
Find file (in list) was just displayed as "Find" with the preview icon,
mostly related to searching.
(Find file is a variation to the filter.)
(To search the files in the commit Find in the FileViewer must be used,
by default Ctrl-F will work too).

Similar change in Compare form.
Added mnemonics
Also removed the hardcoded shortcut that hide the FileViewer shortcut.

Added the preview (magnifier) icon to FileViewer find.

Renamed the "Find file" in Filetree to align with the other uses.
Added mnemonics