git-word-diff navigation and line index marking
Guess added, removed and other (mixed) lines for
git-word-diff from the detected coloring.
This is a heuristic that covers at least the basic usage.

Refactor next/previous section to use the "line analyzis".
This simplifies the implementation, special handling for grep and
range-diff can be removed.

The behavior for git-range-diff is aligned with other diff presentations
as next/previous goes to next header block instead of next header as
before.
range-diff also show line numbers.