Fix theme migration bug (#11629)
Migrating the theme from settings that does not include any color overrides resulted in a migrated theme with only black in it.

This was due to a bug in SettingsSource.GetColor which would return Color.Empty (which is black) for unset colors instead of returning the provided default color. ColorTranslator.FromHtml returns Color.Empty if the input is null or empty.