Difftastic as diff viewer
Support Difftastic as an alternative difftool output
in addition to the standard "patch" and recently added
"git-word-diff" views.
The Difftastic output is mostly presented as-is without changes
(except that there is an attempt to apply Git theme colors).
The info is parsed to show file line info and allow navigation.