fix: label "No changes" to no more hiding first file of the list
when opening "FormLog"
even if it was set as not visible.

Fixes #11679

Reproduce step: double click on a revision in the revision grid

Regression introduced by f8c035302bba1351d88b5159e5ec93bfbbc5933d

/!\ Technical reason of the bug:
Previous code was:

            FilterComboBox.Visible = filesPresent || (SearchComboBox.Visible && !string.IsNullOrEmpty(SearchComboBox.Text));
            NoFiles.Visible = !FilterComboBox.Visible;

But as `FilterComboBox.Visible` is not just a property storing the value directly to a backed field
(it has a complex WinFoms internal logic),
 the value retrieved by the getter ( to after set `NoFiles.Visible`) is sometimes wrong and not the one set to the property :(
The value is not "well" set due to complex internal logic and so a wrong value is used after.

Using a temporary variable fix the issue because the set of the 2 properties has no more impact on each others!