feat: GE theme ANSI terminal colors
Adjust the ANSI terminal colors to align with the GE theme.
The background/foreground colors are separate to improve the
display for the extremely bright GE colors.

The theme is somehow based on mintty Helmhotz, but heavily modified
especially for background (reverse video) to align with the
existing GE theme colors.