amend: Keep linenumber when switching viewer
Try to keep the linenumber in the fileviewer if the identification
for the contents is the same as the captured identification
(using the filename as identification).

Previously the line was restored if the number of lines were the same
or if it was a diff and the first line matched.
Some situations where this did not work:
* A new unrelated file with the same number of lines.
* Switch diffviewer to/from difftastic (as header differs)
* Select a new commit for a file in the file tree,
where the number of lines differed in the other commit.