amend! fix: RunCommand() should not throw when exit is checked
Some methods intend to check the Git output, but the
command always throws at exit.
throwOnErrorExit need to be configurable for these methods.

This occurs for instance when checking the characters in a
branch name and staging files.

Avoid echoing the branchname when creating a branch
(as output in RunCommand() is written to the console).