feat: Cache diff highlighting
Cache the diff markings, only calculate when a file is added,
not everytime marking is changed in the file.
This was already the case for diff coloring by Git,
but not by the (simpler) GE engine.
The caching also applies to the inline difference calculations,
to dim part of lines that are unchanged.