fix: Reset current ansi terminal color
An empty ansi terminal sequence must be handled as an explicit reset.
This could cause that foreground colors were applied when black was to be used.

Adjust overridden colors for range-diff.
contextBold should not be reverse bold and
newDimmed/oldDimmed should be reverse.
(There is a minimal difference bold/normal black, but this context is
not very interesting.)