feat: branch diff adjust icons for  file status
Modify the icons for the file status list when comparing branches
with a common base, the info with feedback for where the files
are changed: Unequal/B (second), A (first), Same (both changes the same).

Color is added to give some feedback of the change in addition to the
(pixel) character: B (blue-violet), A (yellow) and same (green)
(unequal was red already).

Also prepare the icons for dark mode.

Note that only Modified is changed, not added/removed/copied/renamed,
pending feedback (all submodule related are not changed, as before).