feat: support for dark app mode
Basic adaption of .net9 dark app mode.
The themeing is adopted to use Windows dar app mode enabled in .net9.
The theming engine does not allow override of any system colors.

The GE 3.x dark theme are adjusted to
be mostly working with SystemColorMode.Dark.
This is incomplete, a number of overrides and tuning still needed.