fix(theme): Remove unused hook methods (#12127)
Remove the EasyHook handling that was required for theming in
GE v3.x with .NET framework v4. This solution has has not
worked since GE v4, that use .NET core (.net6 and later).
.net9 has experimental support for dark mode that can enable
the use of dark mode.

The activation of hooking was disabled by the define SUPPORT_THEME_HOOKS.
A number of unused methods was used too,
as well as the EasyHooks submodule.

The .net9 solution do not allow customization of system colors,
so overrides were removed from the themes.
As the darksilver theme mostly changed system colors, it was removed.

The invariant theme system colors is required by themeing to adjust
colors and still requires the system colors.
The color for MenuHighlight was corrected in the invariant theme.

ButtonFace, ButtonHighlight, ButtonShadow was assumed to be
duplicates of Control* colors, this was not correct and the handling
was removed.

The workaround for dark header was removed too.