style: Darker panel background
Change to make the GE default theme changes the panel backgrounds
not so extremely bright, more like Visual Studio Code "Modern Light"
theme.
The GitExtensionsDialog are updated too but not the old style
GitExtensionsForm.
The grid view and editors are unchanged.

The line number panel is made darker to contrast to
the text and panels.