fixup! fix: read worktree files from the filesystem
for worktree the TreeGuid is only valid if the file is not dirty.
The treeId is now included for worktree files but must not be used when
viewing files, must always get from the file system.

Similar, for index files the treeid is not immutable and must be
reevaluated.

Regression from 1bb9dad31922f7557acddabaed5de1faac24d7df