amend! fix: read worktree files from the filesystem
fix: read worktree files from the filesystem

For index, the TreeGuid is mutable and must be refreshed every time it is used.
For worktree, the TreeGid is metadata but must not be used to to view files etc.
(if not dirty, the TreeGuid is for Index and could be OK,
that would be more complex than accessing the filesystem).

When viewing files, regression from 1bb9dad31922f7557acddabaed5de1faac24d7df
The TreeGuid is also used when linepatching, that usage was