fix: no throw for commit count
If a commit do not exist in the arguments to 'rev-list --left-count',
the command should not throw.
For submodules, the old/new commit may not exist anymore.
When comparing branches, both commits must exist but this is
common code in IGitModule so it should not be changed now.

Before b6fb8006d338cd435a69ef14e61a443e06d97580
the implementation was separate with throw suppressed for
submodules but enabled for branch compare.

Also suppress cache when comparing branches.