fix: read worktree files from the file system (#12496)
For index, the TreeId (blobId) is mutable and must be refreshed every time it is used.
For worktree, the TreeId is metadata but must not be used to to view files etc.
(if not dirty, the TreeId is for Index and could be OK,
that would be more complex than reading from the file system).

When viewing files, regression from 1bb9dad31922f7557acddabaed5de1faac24d7df
The TreeId is also used when linepatching, that usage was
not updating the treeId before either.