fix: Theme loading error suppressions
* Try loading a color from the invariant theme if not found in the
current theme.
The explicit '@import url("invariant.css");' in user themes is then not
required.

* Ignore the occurrence of color "Graph" in themes.
The color was removed in 5.2.1 -> 6.0 and the themes are not
explicitly migrated removing the user color.

* Limit the printout when a color cannot be parsed.
The full traceback intimidates the user, the color should be
printed on the fist line.
The full traceback is still written to the trace.