fix: ls-files %objecttype Git version check (#12599)
Update the version check as %objecttype was introduced later than the
--format option.
Follow-up to #12501

Note: The actual version in which %objecttype is supported
is in 2.42, this reuses another constant for 2.46
to avoid breaking the Extensibility.Git version, so the fix is used in
a few version not really needed.

Update the minimal version that is somehow supported by
usage to 2.43, used in Ubuntu 22.04.